Added the API routes for students_assignments to the server files
